+++ src/display.h
@@ -25,7 +25,7 @@ public:
 	virtual bool handle_events(Emulator &emu) = 0;
 	virtual uint8_t get_keystate() = 0;
 
-	virtual void play_audio(const uint8_t* samples) = 0;
+	virtual void play_audio(const float* samples) = 0;
 };
 
 #endif
