+++ src/emulator.h
@@ -46,7 +46,7 @@ public:
 	const uint32_t* get_pixel_data() const;
 	const std::string get_game_name() const;
 
-	const uint8_t* get_audio_data() const;
+	const float* get_audio_data() const;
 	void clear_audio();
 
 	void save() const;
