+++ src/sdl_display.h
@@ -45,7 +45,7 @@ public:
 	bool handle_events(Emulator &emu);
 	uint8_t get_keystate();
 
-	void play_audio(const uint8_t *samples);
+	void play_audio(const float *samples);
 };
 
 #endif
